design corrections

+++ src/processor.py
@@ -10,7 +10,7 @@ class Processor:
         self.bots = []
 
     def add_bot(self, new_bot_stats: tuple) -> None:
-        """Adds a bot to the processor list, if it not in list."""
+        """Adds a bot to the proc list, if it not in list."""
         for bot in self.bots:
             if bot.get_number() == new_bot_stats[0]:
                 return
@@ -18,8 +18,17 @@ class Processor:
                       new_bot_stats[3], new_bot_stats[4])
         self.bots.append(new_bot)
 
+    def add_all_bots(self) -> None:
+        """Adds all bot to the proc's list."""
+        conn = sqlite3.connect(r'resources/bots.db')
+        cur = conn.cursor()
+        cur.execute('''SELECT * FROM bots''')
+        all_bots_stats = cur.fetchall()
+        for bot_stats in all_bots_stats:
+            self.add_bot(bot_stats)
+
     def delete_bot(self, delete_bot_number: int) -> None:
-        """Deletes bot from the processor list."""
+        """Deletes bot from the proc list."""
         delete_bot = None
         for bot in self.bots:
             if bot.get_number() == delete_bot_number:
@@ -28,7 +37,6 @@ class Processor:
         if delete_bot is None:
             return
         self.bots.remove(delete_bot)
-        return
 
     def create_bot_db(self, x: int, y: int) -> dict:
         """Creates new bot."""
@@ -51,6 +59,13 @@ class Processor:
         """Deletes bot with given number."""
         conn = sqlite3.connect(r'resources/bots.db')
         cur = conn.cursor()
+        cur.execute('''SELECT * FROM bots WHERE id = ? ORDER BY id DESC LIMIT 1 ''', (bot_number,))
+        deleting_bot_stats = cur.fetchone()
+        if not deleting_bot_stats:
+            return {'bot': 'does not exists'}
+        deleting_bot = self.get_current_bot_instance(bot_number)
+        if deleting_bot.is_moving:
+            return {'ERROR': 'You cannot bot while its moving'}
         self.delete_bot(bot_number)
         cur.execute('''DELETE FROM bots WHERE id = ?''', (bot_number,))
         conn.commit()
@@ -60,12 +75,15 @@ class Processor:
         """Deletes all bots from the database."""
         conn = sqlite3.connect(r'resources/bots.db')
         cur = conn.cursor()
+        for bot in self.bots:
+            if bot.is_moving:
+                return {'ERROR': 'You cannot delete all bots, while at least one moving'}
         self.bots.clear()
         cur.execute('''DELETE from bots''')
         conn.commit()
         return {'list of bots': 'is empty'}
 
-    def get_current_bot(self, searched_bot_number: int):
+    def get_current_bot_instance(self, searched_bot_number: int):
         """Returns bot instance with a searched bot number."""
         for bot in self.bots:
             if bot.get_number() == searched_bot_number:
@@ -115,13 +133,11 @@ class Processor:
         new_targets = (bot.get_target_x(), bot.get_target_y(), bot.get_number())
         cur.execute("""UPDATE bots SET target_x = ?, target_y = ? WHERE id = ?""", new_targets)
 
-        movement_is_over = False
-        while movement_is_over is False:  # bad decision, but i don't come up another yet
+        while not bot.get_x() == bot.get_target_x() and not bot.get_y() == bot.get_target_y():
             self.bot_step_x(bot)
             self.bot_step_y(bot)
             new_coordinates = (bot.get_x(), bot.get_y(), bot.get_number())
             cur.execute("""UPDATE bots SET x = ?, y = ? WHERE id = ?""", new_coordinates)
             conn.commit()
             time.sleep(1)
-            if bot.get_x() == bot.get_target_x() and bot.get_y() == bot.get_target_y():
-                return
+
